requirements.txt
opencv-python
numpy
mediapipe==0.10.33
twilio==9.0.4
PyYAML==6.0.1
pytest==8.1.1
